+++ relm4-macros/src/funcs.rs
@@ -5,6 +5,8 @@ use syn::{spanned::Spanned, Error, ImplItemMethod, Result};
 pub(super) struct Funcs {
     pub pre_init: Option<TokenStream2>,
     pub post_init: Option<TokenStream2>,
+    pub pre_connect_components: Option<TokenStream2>,
+    pub post_connect_components: Option<TokenStream2>,
     pub manual_view: Option<TokenStream2>,
 }
 
@@ -12,6 +14,8 @@ impl Funcs {
     pub fn new(funcs: &[ImplItemMethod]) -> Result<Self> {
         let mut pre_init = None;
         let mut post_init = None;
+        let mut pre_connect_components = None;
+        let mut post_connect_components = None;
         let mut manual_view = None;
 
         for func in funcs {
@@ -23,7 +27,7 @@ impl Funcs {
                 if pre_init.is_some() {
                     return Err(Error::new(
                         func.span().unwrap().into(),
-                        "widget method defined multiple times",
+                        "pre_init method defined multiple times",
                     ));
                 }
                 pre_init = Some(tokens);
@@ -35,6 +39,22 @@ impl Funcs {
                     ));
                 }
                 post_init = Some(tokens);
+            } else if ident == "pre_connect_components" {
+                if pre_connect_components.is_some() {
+                    return Err(Error::new(
+                        func.span().unwrap().into(),
+                        "pre_connect_components method defined multiple times",
+                    ));
+                }
+                pre_connect_components= Some(tokens);
+            } else if ident == "post_connect_components" {
+                if post_connect_components.is_some() {
+                    return Err(Error::new(
+                        func.span().unwrap().into(),
+                        "post_connect_components method defined multiple times",
+                    ));
+                }
+                post_connect_components = Some(tokens);
             } else if ident == "manual_view" {
                 if manual_view.is_some() {
                     return Err(Error::new(
@@ -54,6 +74,8 @@ impl Funcs {
         Ok(Funcs {
             pre_init,
             post_init,
+            pre_connect_components,
+            post_connect_components,
             manual_view,
         })
     }